+++ src/users/roles.decorator.ts
@@ -1,5 +1,4 @@
 import { SetMetadata } from '@nestjs/common';
-import { UserRole } from './user.entity';
-
+import { UserRole } from './entities/user-role';
 export const ROLES_KEY = 'roles';
 export const Roles = (...roles: UserRole[]) => SetMetadata(ROLES_KEY, roles);
