+++ src/users/users.controller.ts
@@ -1,15 +1,13 @@
-import { UsersService } from './users.service';
 import { Body, Controller, Get, Post } from '@nestjs/common';
-import {
-  AuthenticateUserDTO,
-  CreateUserDTO,
-  UserDetail,
-  UsersResponse,
-} from './users.dto';
-import { User } from './user.decorator';
-import { UserEntity, UserRole } from './user.entity';
-import { Roles } from './roles.decorator';
 import { Public } from 'src/auth/Public.decorator';
+import { UserRole } from './entities/user-role';
+import { UserDetail } from './entities/user.detail';
+import { UserEntity } from './entities/user.entity';
+import { UserPublic } from './entities/user.public';
+import { Roles } from './roles.decorator';
+import { User } from './user.decorator';
+import { AuthDTO, CreateUserDTO, UsersResponse } from './users.dto';
+import { UsersService } from './users.service';
 
 @Controller('users')
 export class UsersController {
@@ -20,10 +18,9 @@ export class UsersController {
   async find(): Promise<UsersResponse> {
     return {
       total: await this.usersService.count(),
-      users: await this.usersService.find(),
+      users: await this.usersService.find({ cls: UserPublic }),
     };
   }
-
   @Get('profile')
   async profile(@User() user: UserDetail): Promise<UserDetail> {
     return user;
@@ -35,11 +32,8 @@ export class UsersController {
     return this.usersService.create(body);
   }
 
-  /**
-   * 자기 자신만 바꿀 수 있다. 로그인 한 정보가 아닌 body정보로 '나'를 인증
-   */
   @Post('delete')
-  delete(@Body() body: AuthenticateUserDTO): Promise<UserEntity> {
-    return this.usersService.delete(body);
+  delete(@Body() body: AuthDTO): Promise<UserDetail> {
+    return this.usersService.deleteOne(body);
   }
 }
