--- src/users2/roles.decorator.ts
@@ -1,4 +1,0 @@
-import { SetMetadata } from '@nestjs/common';
-import { UserRole } from './entities/user-role';
-export const ROLES_KEY = 'roles';
-export const Roles = (...roles: UserRole[]) => SetMetadata(ROLES_KEY, roles);
