--- src/users2/user.decorator.ts
@@ -1,8 +1,0 @@
-import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-
-export const User = createParamDecorator(
-  (data: unknown, ctx: ExecutionContext) => {
-    const request = ctx.switchToHttp().getRequest();
-    return request.user;
-  },
-);
