--- src/users2/users.module.ts
@@ -1,23 +1,0 @@
-import { APP_GUARD } from '@nestjs/core';
-import { Module } from '@nestjs/common';
-import { MongooseModule } from '@nestjs/mongoose';
-import { UsersRepository } from './users.repository';
-import { UserSchemaProvider } from './entities/user.entity';
-import { RolesGuard } from './roles.guard';
-import { UsersService } from './users.service';
-import { UsersController } from './users.controller';
-
-@Module({
-  imports: [MongooseModule.forFeatureAsync([UserSchemaProvider])],
-  providers: [
-    UsersRepository,
-    UsersService,
-    {
-      provide: APP_GUARD,
-      useClass: RolesGuard,
-    },
-  ],
-  controllers: [UsersController],
-  exports: [UsersService],
-})
-export class UsersModule {}
